--- frontend/src/Modal.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './css/Modal.css';
-
-const Modal = ({ isOpen, children, onClose }) => {
-  if (!isOpen) return null;
-
-  return (
-    <div className="modal-overlay" onClick={onClose}>
-      <div className="modal-content" onClick={e => e.stopPropagation()}>
-        {children}
-      </div>
-    </div>
-  );
-};
-
-export default Modal;
