+++ frontend/src/components/loginAndRegister/ForgotPassword.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
-import logo from './assets/FinalLogo.png';
-import { instance } from './App';
-import './css/ForgotPassword.css';
+import logo from '../../assets/FinalLogo.png';
+import { instance } from '../../App';
+import '../../css/ForgotPassword.css';
 
 const ForgotPassword = ({ onClose, onSwitchForm }) => {
 	const navigate = useNavigate();
