+++ frontend/src/components/pageFeatures/MainContent.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import AnimeCard from './AnimeCard'
-import './css/MainContent.css'
+import AnimeCard from '../animeCards/AnimeCard'
+import '../../css/MainContent.css'
 
 function MainContent(props) {
     return (
