+++ frontend/src/components/pageFeatures/Modal.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import '../../css/Modal.css';
+
+const Modal = ({ isOpen, children, onClose }) => {
+  if (!isOpen) return null;
+
+  return (
+    <div className="modal-overlay" onClick={onClose}>
+      <div className="modal-content" onClick={e => e.stopPropagation()}>
+        {children}
+      </div>
+    </div>
+  );
+};
+
+export default Modal;
