+++ frontend/src/css/HomePage.css
@@ -1,0 +1,175 @@
+@import url('https://fonts.googleapis.com/css2?family=Bangers&display=swap');
+
+* {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
+}
+
+body {
+    background-color: #282c34;
+}
+
+#background-video {
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    min-width: 100%;
+    min-height: 100%;
+    width: auto;
+    height: auto;
+    z-index: -1;
+}
+
+.home-page {
+    width: 100%;
+    height: 80px;
+    background-color: red;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    position: fixed;
+    top: 0;
+}
+
+.container {
+    display: flex;
+    justify-content: center;
+    margin-top: 0;
+    align-items: flex-start;
+    width: 100%;
+}
+
+.anime-row {
+    display: flex;
+    flex-direction: column;
+    overflow-y: auto;
+    margin-top: 0;
+    padding-top: 0;
+    height: calc(100vh - 80px);
+}
+
+.row {
+    display: flex;
+    gap: 20px;
+    flex-wrap: wrap;
+    justify-content: center;
+}
+
+.anime-row::-webkit-scrollbar {
+    display: none;
+}
+
+.card {
+    width: 190px;
+    height: 260px;
+    padding: 0 0 1rem 0;
+    box-shadow: 5px 5px 5px rgba(0,0,0,0.3);
+    text-align: center;
+    overflow: hidden;
+    position: relative;
+    max-width: 190px;
+}
+
+.card img {
+    width: 100%;
+    height: 100%;
+}
+
+.logo-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.logo-container img {
+    height: 250px;
+    width: auto;
+    margin-bottom: 10px;
+}
+
+.logo-and-form-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    height: 100vh;
+    margin-top: -80px;
+    padding-top: 80px;
+}
+
+.buttons {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    gap: 10px;
+}
+
+.button {
+    padding: 15px 30px;
+    font-size: 1.5rem;
+    font-family: 'Bangers', cursive;
+    background-color: #2E77AE;
+    border-image: linear-gradient(to right, #2196F3, #f44336) 1;
+    border: 10px solid;
+    color: #fff;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.button:hover {
+    background-color: #0d2137;
+}
+
+.about-us-button {
+    padding: 10px 20px;
+    font-size: 16px;
+    font-family: 'Bangers', cursive;
+    background-color: #2E77AE;
+    color: #fff;
+    border: none;
+    border-radius: 10px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+}
+
+.about-us-button:hover {
+    background-color: #0d2137;
+}
+
+.topbar {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+    position: fixed;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 80px;
+    background-color: #111920;
+    z-index: 2;
+    padding: 0 20px;
+}
+
+@keyframes scrollUp {
+    0% {
+      transform: translateY(100%);
+    }
+    100% {
+      transform: translateY(-100%);
+    }
+  }
+  
+  .scroll-animation {
+    animation: scrollUp 60s linear infinite;
+  }
+
+  html {
+    overflow-y: scroll;
+}
+
+.form-active {
+    align-items: flex-start;
+    padding-top: 100px;
+}
